Make webpack-config files more clear
Why:

* #229, notation from @cpylua

How:

* peel babel-loader scss-loader postcss-loader out, reunion as a new file.
